Extract Puzzle component #2
